feat: Counties (list, add, edit, recovery, bulk-actions) | feat: Cities (list, add, edit, recovery, bulk-actions)

+++ app/Models/Country.php
@@ -3,15 +3,14 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Support\Facades\Storage;
 
 class Country extends Model
 {
-    protected $fillable = ['name', 'continent_id', 'img_path', 'short_desc', 'slug', 'show_on_homepage'];
+    use SoftDeletes;
 
-    public function continent()
-    {
-        return $this->belongsTo(Continent::class);
-    }
+    protected $guarded = ['id', 'created_at', 'updated_at'];
 
     public function cities()
     {
@@ -25,15 +24,35 @@ class Country extends Model
         })->where('tours.is_active', true);
     }
 
+    public function seo()
+    {
+        return $this->morphOne(Seo::class, 'seoable');
+    }
+
     protected static function boot()
     {
         parent::boot();
 
-        static::deleting(function ($country) {
-            // Delete all related cities when the country is deleted
-            $country->cities()->each(function ($city) {
-                $city->delete();
-            });
+        static::deleting(function ($item) {
+            if ($item->isForceDeleting()) {
+                self::deleteImage($item->featured_image);
+                if ($item->seo) {
+                    self::deleteImage($item->seo->seo_featured_image);
+                    self::deleteImage($item->seo->fb_featured_image);
+                    self::deleteImage($item->seo->tw_featured_image);
+                    $item->seo->delete();
+                }
+                $item->cities()->each(function ($city) {
+                    $city->delete();
+                });
+            }
         });
+    }
+
+    public static function deleteImage($path)
+    {
+        if ($path && Storage::disk('public')->exists($path)) {
+            Storage::disk('public')->delete($path);
+        }
     }
 }
